pyproject.toml
[project]
name = "arquive-adquisition"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "imutils>=0.5.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "uvicorn[standard]>=0.34.2",
]
